+++ rollbar/__init__.py
@@ -118,7 +118,7 @@ log = logging.getLogger(__name__)
 
 agent_log = None
 
-VERSION = '0.7.5'
+VERSION = '0.7.6'
 DEFAULT_ENDPOINT = 'https://api.rollbar.com/api/1/'
 DEFAULT_TIMEOUT = 3